added title scene, updated ui animated zombies in title scene added music
Added a title scene that allows users to view credits and controls as well as  start the game. All UI is updated to scale with screen size and has good fonts and all have been cleaned up a bit. Started working on animating the zombies. Added a song to the title scene and the main scene but need to do some stuff to the main song